rag_project/rag_system/requirements.txt
faiss-cpu
numpy
requests
torch
transformers